# Optimize voctree build by up to 40 times for certain problems