This PR fixes 3 performance problems in voctree which cause inefficiency for relatively small problems:
 - all center updates depend on a single lock during clustering
 - thread pool is created per-iteration during clustering
 - thread pool is created per-iteration during k-means initialization

This PR disables parallelism in case problem size is small and introduces a lock for each cluster center to help when this is not the case. Additionally, k-means initialization now runs in parallel for trials which reduces the thread fanout when calculating new distance sum.

I haven't tested this with large problems, but at least vocabularyTreeBuild test became faster by up to 40 times on a machine with AMD 2990WX (~4s -> 0.1s). The runtime of another relevant test `voctree_kmeans` has been improved from ~3.7s to ~2.7s.

Note that AMD 2990WX has 32 cores/64 threads, so any lock contention problems are exacerbated there. It's not top of the line machine though, as 128 core/256 thread machines have been available for several years already.