@p12tic 2022-10-10
A further optimization pushed the improvement 10 -> 40 times. I've edited the PR description to reflect this.